+++ core/components/minishop3/elements/tasks/csvImport.php
@@ -1,14 +1,15 @@
 <?php
 
-use MiniShop3\Utils\ImportCSV;
+use MiniShop3\Services\Product\Import\ProductImportService;
 
-/** @var modX $modx */
+/** @var \MODX\Revolution\modX $modx */
 /** @var sFileTask $task */
 /** @var sTaskRun $run */
 /** @var array $scriptProperties */
 
-$importCSV = new ImportCSV($modx);
-$result = $importCSV->process($scriptProperties);
+/** @var ProductImportService $importService */
+$importService = $modx->services->get('ms3_product_import');
+$result = $importService->process($scriptProperties);
 
 if (!is_array($result) || empty($result['success'])) {
     $errorMessage = is_array($result) && !empty($result['message'])
